// src/generated/prisma/enums.ts
export const Role = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  PRESIDENT: 'PRESIDENT',
  SECRETARY: 'SECRETARY',
  DIRECTOR: 'DIRECTOR',
  BOARD_MEMBER: 'BOARD_MEMBER',
  VIEWER: 'VIEWER'
} as const

export type Role = (typeof Role)[keyof typeof Role]